feat(usercase) WIP: category and product)

+++ backend/core/product-service/src/main/java/com/github/sfidencio/adapter/output/database/entities/ProductEntity.java
@@ -4,11 +4,13 @@ package com.github.sfidencio.adapter.output.database.entities;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.MongoId;
 
+import java.math.BigDecimal;
+
 @Document(collection = "products")
 public class ProductEntity {
     @MongoId
     private String id;
     private String description;
-    private double price;
-    private double stock;
+    private BigDecimal price;
+    private BigDecimal stock;
 }
